feat: add authorization ownership check in controllers

+++ app/Http/Controllers/TransactionController.php
@@ -2,8 +2,6 @@
 
 namespace App\Http\Controllers;
 
-use App\Http\Requests\Transaction\TransactionStoreRequest;
-use App\Http\Requests\Transaction\TransactionUpdateRequest;
 use Request;
 use App\Models\Transaction;
 use App\QueryOptions\Sort\Date;
@@ -11,7 +9,10 @@ use App\QueryOptions\Filter\Type;
 use App\QueryOptions\Sort\Amount;
 use App\Models\TransactionCategory;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Pipeline;
+use App\Http\Requests\Transaction\TransactionStoreRequest;
+use App\Http\Requests\Transaction\TransactionUpdateRequest;
 
 class TransactionController extends Controller
 {
@@ -29,7 +30,7 @@ class TransactionController extends Controller
                 ->back()
                 ->withErrors([
                     'transaction_category' =>
-                    'Invalid transaction category or type.',
+                        'Invalid transaction category or type.',
                 ]);
         }
 
@@ -52,6 +53,8 @@ class TransactionController extends Controller
 
     public function show(Transaction $transaction)
     {
+        Gate::authorize('view', $transaction);
+
         return view('transactions.show', [
             'transaction' => $transaction,
         ]);
@@ -61,6 +64,8 @@ class TransactionController extends Controller
         TransactionUpdateRequest $request,
         Transaction $transaction
     ) {
+        Gate::authorize('update', $transaction);
+
         $isTransactionValid = TransactionCategory::query()
             ->where('id', '=', $request['transaction_category_id'])
             ->where('transaction_type', '=', $request['type'])
@@ -71,7 +76,7 @@ class TransactionController extends Controller
                 ->back()
                 ->withErrors([
                     'transaction_category' =>
-                    'Invalid transaction category or type.',
+                        'Invalid transaction category or type.',
                 ]);
         }
 
@@ -84,6 +89,8 @@ class TransactionController extends Controller
 
     public function destroy(Transaction $transaction)
     {
+        Gate::authorize('delete', $transaction);
+
         $transaction->delete();
 
         return redirect()->back();
